requirements.txt
aiodns==2.0.0
aiofiles==0.6.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hstspreload==2020.12.22
httpcore==0.13.2
httptools==0.2.0
httpx==0.18.1
hyperframe==6.0.1
idna==2.9
multidict==5.1.0
pycares==3.1.1
pycparser==2.20
python-dotenv==0.17.1
requests==2.25.1
rfc3986==1.4.0
sanic==21.3.4
sanic-routing==0.6.2
sniffio==1.2.0
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
websockets==8.1
yarl==1.6.3
